categorize windows.hollowprocesses as malware
